Refactor MessageApi tests
