+++ fe1-web/network/__tests__/MessageApiWitness.test.ts
@@ -1,0 +1,40 @@
+import 'jest-extended';
+import '__tests__/utils/matchers';
+import {
+  ActionType,
+  MessageData,
+  ObjectType,
+  WitnessMessage,
+} from 'model/network/method/message/data';
+import { defaultMessageDataFields, mockLao, mockLaoId } from '__tests__/utils/TestUtils';
+import { Base64UrlData } from 'model/objects';
+import { publishMock } from '__mocks__/publishMock';
+import { OpenedLaoStore } from 'store';
+import * as msApi from '../MessageApi';
+
+const checkDataWitnessMessage = (obj: MessageData) => {
+  expect(obj.object).toBe(ObjectType.MESSAGE);
+  expect(obj.action).toBe(ActionType.WITNESS);
+
+  const data: WitnessMessage = obj as WitnessMessage;
+  expect(data).toContainKeys([...defaultMessageDataFields, 'message_id', 'signature']);
+  expect(data.message_id).toBeBase64Url();
+  expect(data.signature).toBeBase64Url();
+};
+
+beforeEach(() => {
+  OpenedLaoStore.store(mockLao);
+  publishMock.mockClear();
+});
+
+describe('MessageApi', () => {
+  it('should create the correct request for requestWitnessMessage', async () => {
+    await msApi.requestWitnessMessage(`/root/${mockLaoId}`,
+      Base64UrlData.encode('randomMessageId'));
+
+    expect(publishMock).toBeCalledTimes(1);
+    const [channel, msgData] = publishMock.mock.calls[0];
+    expect(channel).toBe(`/root/${mockLaoId}`);
+    checkDataWitnessMessage(msgData);
+  });
+});
